+++ tests/end-to-end/home.test.ts
@@ -52,12 +52,10 @@ describeEndToEndTest('Home recycling', () => {
     });
 
     const narrowAccessSchemeText = page
-      .getByText(
-        mockedLaResponse.properties[PROPERTY_TYPE_EN.NARROW_ACCESS][0].name,
-      )
+      .getByText(PROPERTY_TYPE_EN.NARROW_ACCESS)
       .first();
     const kerbsideSchemeText = page
-      .getByText(mockedLaResponse.properties[PROPERTY_TYPE_EN.KERBSIDE][0].name)
+      .getByText(PROPERTY_TYPE_EN.KERBSIDE)
       .first();
 
     await widget.evaluate((node) =>
@@ -135,9 +133,7 @@ describeEndToEndTest('Home recycling', () => {
     });
 
     const kerbsideSchemeText = page
-      .getByText(
-        LocalAuthorityResponse.properties[PROPERTY_TYPE_EN.KERBSIDE][0].name,
-      )
+      .getByText(PROPERTY_TYPE_EN.KERBSIDE)
       .first();
     const kerbsideSchemeLink = page
       .locator('a', { has: kerbsideSchemeText })